# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DOCSWAP is a platform designed for realtors to efficiently share, manage, and analyze documents on a neighborhood level. It enables realtors to make better use of the content created for each listing by organizing it into a searchable database that supports both uploads and downloads of documents. These documents are categorized to facilitate market analysis, putting the power of data directly into the hands of realtors and market analysts interested in housing information in the lower mainland.",
        "title": "DOCSWAP API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "DOC-SWAP Support",
            "url": "http://www.doc-swap.com/support",
            "email": "support@doc-swap.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/userdocument/": {
            "get": {
                "description": "Get all user document records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userdocuments"
                ],
                "summary": "Retrieve all user document records",
                "parameters": [
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include soft deleted user documents",
                        "name": "includeDeleted",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include full user and document details",
                        "name": "full",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of user document records",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserDocument"
                            }
                        }
                    },
                    "400": {
                        "description": "Error: No user document records found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user document record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userdocuments"
                ],
                "summary": "Create a new user document record",
                "parameters": [
                    {
                        "description": "User Document object that needs to be created",
                        "name": "userDocument",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserDocument"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created the user document record",
                        "schema": {
                            "$ref": "#/definitions/models.UserDocument"
                        }
                    },
                    "400": {
                        "description": "Error: Unable to create user document record",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/userdocument/bulk": {
            "post": {
                "description": "Create multiple new user document records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userdocuments"
                ],
                "summary": "Create multiple new user document records",
                "parameters": [
                    {
                        "description": "User Document objects that need to be created",
                        "name": "userDocuments",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserDocument"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created the user document records",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserDocument"
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Unable to create user document records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete multiple user document records by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userdocuments"
                ],
                "summary": "Delete multiple user document records by their ID",
                "parameters": [
                    {
                        "description": "User Document objects that need to be deleted",
                        "name": "userDocuments",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserDocument"
                            }
                        }
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Set to false to fully delete the user document records",
                        "name": "softDelete",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted the user document records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error: Unable to delete user document records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/userdocument/{userId}/{documentId}": {
            "get": {
                "description": "Get a user document record by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userdocuments"
                ],
                "summary": "Retrieve a user document record by its ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Document ID",
                        "name": "documentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include soft deleted user documents",
                        "name": "includeDeleted",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include full user and document details",
                        "name": "full",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the user document record",
                        "schema": {
                            "$ref": "#/definitions/models.UserDocument"
                        }
                    },
                    "400": {
                        "description": "Error: No user document record found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a user document record by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userdocuments"
                ],
                "summary": "Delete a user document record by its ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Document ID",
                        "name": "documentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Set to false to fully delete the user document record",
                        "name": "softDelete",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted the user document record",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error: Unable to delete user document record",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Category": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "$ref": "#/definitions/models.Category"
                },
                "parentID": {
                    "type": "integer"
                }
            }
        },
        "models.Config": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Document": {
            "type": "object",
            "properties": {
                "additionalDetails": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "approvedAt": {
                    "type": "string"
                },
                "approvedBy": {
                    "type": "integer"
                },
                "approvedByUser": {
                    "$ref": "#/definitions/models.User"
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "categoryID": {
                    "type": "integer"
                },
                "creditValue": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "fileStorageURL": {
                    "type": "string"
                },
                "fileType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DocumentTag"
                    }
                },
                "uploadedAt": {
                    "type": "string"
                }
            }
        },
        "models.DocumentTag": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "document": {
                    "$ref": "#/definitions/models.Document"
                },
                "documentID": {
                    "type": "integer"
                },
                "tag": {
                    "$ref": "#/definitions/models.Tag"
                },
                "tagID": {
                    "type": "integer"
                }
            }
        },
        "models.Permission": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RolePermission"
                    }
                }
            }
        },
        "models.Role": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RolePermission"
                    }
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserRole"
                    }
                }
            }
        },
        "models.RolePermission": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "permission": {
                    "$ref": "#/definitions/models.Permission"
                },
                "permissionID": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "roleID": {
                    "type": "integer"
                }
            }
        },
        "models.Tag": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DocumentTag"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "authenticationType": {
                    "type": "string"
                },
                "biography": {
                    "type": "string"
                },
                "configs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserConfig"
                    }
                },
                "deletedAt": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserDocument"
                    }
                },
                "email": {
                    "type": "string"
                },
                "externalUserID": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserRole"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserConfig": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/models.Config"
                },
                "configID": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.UserDocument": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "document": {
                    "$ref": "#/definitions/models.Document"
                },
                "documentID": {
                    "type": "integer"
                },
                "isOwner": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.UserRole": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "roleID": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}